+++ src/state/pwallet/hooks/dummy.ts
@@ -1,0 +1,7 @@
+export function useDummyGetUserLP() {
+  return { v2IsLoading: false, allPairs: [], allV2PairsWithLiquidity: [] };
+}
+
+export function useDummyCreatePair() {
+  return null;
+}
